data_loader/pyproject.toml
[tool.poetry]
name = "data-loader"
version = "0.1.0"
description = "Data loader for TEDx Italian Speeches RAG"
authors = ["Luca Ferrario <lucaferrario199@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3"
chromadb = ">=0.5.23"
llama-index-core = ">=0.12.8"
llama-index-embeddings-openai = ">=0.3.1"
llama-index-llms-openai = ">=0.3.12"
llama-index-readers-file = ">=0.4.1"
llama-index-vector-stores-chroma = ">=0.4.1"
loguru = ">=0.7.3"
pysqlite3-binary = ">=0.5.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*.py"]